Undo refactor frame_helper for now
